+++ src/Exceptions/CollectionItemNotFound.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Spatie\CollectionMacros\Exceptions;
+
+use Exception;
+
+class CollectionItemNotFound extends Exception
+{
+}