+++ tests/Macros/EachConsTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Spatie\CollectionMacros\Test;
+namespace Spatie\CollectionMacros\Test\Macros;
 
 use Illuminate\Support\Collection;
+use Spatie\CollectionMacros\Test\TestCase;
 
 class EachConsTest extends TestCase
 {